Move ReservationSuccessModal to shared folder
