+++ app/shared/modals/reservation-success/index.js
@@ -1,0 +1,3 @@
+import ReservationSuccessModalContainer from './ReservationSuccessModalContainer';
+
+export default ReservationSuccessModalContainer;
